BIG update to have a the full architecture covered
- A ton of new structures, ideas, workflows and architecture to make any project work, have the possiblity to use Burst with systems, enable binding a UnityComponent to an OSO (ObservableScriptableObject), Registries to query LinkSO (enables relationships between OSOs), and OSO, access methods to make it cleaner, opening up "services" with lazy loading and instantiation for Pooling, automatic pooling of Runtime created OSO's and LinkSO. Different types of registries for different "domains".

A ton.

I still have to add a bunch to it, like how to instantiate a prefab and setup all scriptableObjects from a template (which is what is held in the prefab) to go through the pooling system and registries.  To come soon... Updates to the node graph soon too: renaming to Static Dependency Graph, and creation of the Execution Trace Graph soon for even better in-depth debugging and clarity!