Massive performance changes + some bugfixes
ObservableRuntimeWatcher used to run on update loops to check if a change was made to an SO and all fields would be marked to loop until we find the one that changed.
Now we have two update modes: Default & HighFrequency
Default is for most usages and event driven, so when a change to an observable is flagged, the Runtime Watcher gets trigger which triggers systems.
HighFrequency is for huge amount of SO's having fields changing multiple times per frames (physics, movement, etc). In that case, the change is not registered in a event driven way (would be a bottleneck) instead, we just create hashes and run a bursted parallel jobs to catch which fields changed via nativeArrays. This means that tens of thousands of Observables could change and there would be negligible impact.
Rewriting this we also had a lot og GC allocs which could sometimes spike out of control and not meet a basic target of 60FPS on GC allocations. These were only spikes but still, we could have 3-5KB GC allocs. We are now down to 77B and those are mostly editor related with allocation of jobs and at runtime will be even closer to 0. CPU times are also at around 5ms, and 3.87ms are actually due to the Editor. The target for 120fps is 8ms so we're really good.
Some fixes in persistence of data have been done too, as well as rewrapping some leftover Debug.Log into ReaCSDebug to be able to filter out debug as much as possible with ReaCS Settings.